Move mapif functions from int_pet.c to mapif.c
